Split out registries.private.h to expose test-only APIs
