+++ src/vcpkg-test/registries.cpp
@@ -5,6 +5,7 @@
 
 #include <vcpkg/configuration.h>
 #include <vcpkg/registries.h>
+#include <vcpkg/registries.private.h>
 
 using namespace vcpkg;
